fist level mst posts (without comments)

+++ tests/mevac_test.py
@@ -33,14 +33,14 @@ class TestMastodon(unittest.TestCase):
             self.assertIsInstance(int(media_post_id), int)
             media_post_ids.append(media_post_id)
 
-        post_ids = self.mst.post_status('Integration Private test', media_post_ids, dry_run=False)
+        post_ids = self.mst.post_fb_status('Integration Private test', media_post_ids, dry_run=False)
         self.assertNotEqual(len(post_ids), 0)
         for post_id in post_ids:
             deleted_post_id = self.mst.delete_entity(post_id)
             self.assertEqual(post_id, deleted_post_id)
 
     def test_hidden_post(self):
-        post_ids = self.mst.post_status('Integration Private test', private=True, dry_run=False)
+        post_ids = self.mst.post_fb_status('Integration Private test', visibility=True, dry_run=False)
         for post_id in post_ids:
             deleted_post_id = self.mst.delete_entity(post_id)
             self.assertEqual(post_id, deleted_post_id)
@@ -51,13 +51,14 @@ class TestFbImporter(unittest.TestCase):
     push_env = PushEnv()
 
     def prepare(self):
-        os.remove(self.load_env.db_file)
+        if os.path.exists(self.load_env.db_file):
+            os.remove(self.load_env.db_file)
 
     def run_importer(self):
         importer = FbImporter(self.load_env)
         self.assertIsInstance(importer._facebook_post_file, str)
         self.assertNotEqual(importer._facebook_post_file, '')
-        importer.load_fb_posts()
+        importer.load_fb_posts(dry_run=False)
         self.assertIsNotNone(self.load_env.stat_fb_posts)
         importer.collect_stat()
         importer.print_stat()
@@ -89,17 +90,23 @@ class TestMstImporter(unittest.TestCase):
         importer.collect_stat()
         importer.print_stat()
 
-    # def run_pusher(self):
-    #     pusher = Pusher(self.load_env, self.push_env)
-    #     pushed_posts = pusher.push_fb_posts()
-    #     for pushed_post in pushed_posts:
-    #         self.assertIsInstance(int(pushed_post), int)
-    #         deleted_post_id = pusher._mst.delete_entity(pushed_post)
-    #         self.assertEqual(pushed_post, deleted_post_id)
-    #
+    def run_pusher(self):
+        pusher = Pusher(self.load_env, self.push_env)
+        pushed_posts = pusher.push_mst_posts(dry_run=False)
+
+        for pushed_post in pushed_posts:
+            self.assertIsInstance(int(pushed_post), int)
+            deleted_post_id = pusher._mst.delete_entity(pushed_post)
+            self.assertEqual(pushed_post, deleted_post_id)
+
+    def prepare(self):
+        if os.path.exists(self.load_env.db_file):
+            os.remove(self.load_env.db_file)
+
     def test_importer(self):
+        self.prepare()
         self.run_importer()
-        # self.run_pusher()
+        self.run_pusher()
 
 
 if __name__ == '__main__':
